# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0
paths: 
  /tasks:
    get:
      summary: Get all tasks
      tags: 
        - tasks
      responses: 
        '200':
          description: A list of tasks
          content: 
            application/json: 
              schema: 
                type: array
                items: 
                  $ref: '#/components/schemas/Task'
    post: 
      summary: Create a new task
      tags: 
        - tasks
      requestBody:
        description: The task to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '201':
          description: The created task
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  /tasks/{taskId}:
    patch:
      summary: Update a task
      tags:
        - tasks
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Fields to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '200':
          description: Updated task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    delete:
      summary: Delete a task
      tags:
        - tasks
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted the task
  /tasks/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
    post:
      summary: Create task by user id
      tags:
        - tasks
      
      requestBody:
        description: The user to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '201':
          description: The task create by user id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  /users:
    get:
      summary: Get all users
      tags:
        - users
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Create a new user
      tags:
        - users
      requestBody:
        description: The user to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: The created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{userId}:
    patch:
      summary: Update a user
      tags:
        - users
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Fields to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          decsription: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      summary: Delete a user
      tags:
        - users
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted the user
  /users/{userId}/tasks:
    get:
      summary: Get all users
      tags:
        - tasks
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        password:
          type: string
    Task:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        is_done:
          type: boolean
        user_id:
          type: string